Preliminary URL processing in Trails Controller.
I added the preliminary business logic in the trials controller to process the URLs that
are passed to it. Currently we extract the host, the title, the description. Date processing
is not handled as yet because it is complicated. The same can be said about the
thumbnail generation - which is also absent at this time.