Implemented simple image extraction - extraction algorithm still not perfect!
I have completed the first attempt at extracting the images from a url and presenting a simple UI to allow
users to switch between them. This currently relies of certian HTML properties to be set (namely height
and width) which are not always present. In the future we need a better way of determining good "images"
to use.